feat : 프로젝트 댓글 CRUD 구현 / 프로젝트 별점 구현

+++ src/main/java/cocodas/prier/project/project/Project.java
@@ -147,7 +147,7 @@ public class Project {
         this.teamMate = teamMate;
     }
 
-    public void updateScore(int score) {
+    public void updateScore(Float score) {
         this.score += score;
     }
 }
